Add files via upload
The purpose of this assignment is to create a webpage that extracts dynamic data – i.e. that which is regularly changing – from the Internet and presents it to the user nicely organized and formatted. Examples of such information might include:

1.	The current world population 
2.	The Dow Jones Industrial Average
3.	The present temperature in a given zip-code
4.	The number of Google searches conducted today
5.	The U.S.  national debt
6.	The barrels of oil  pumped today
7.	The liters of wine produced today
8.	The cans of Coca Cola drunk today in the U.S.
9.	The number of text messages sent today
10.	The number of Facebook posts today

Your page should include at least five such statistics (you can choose different one instead) obtained from as many distinct websites. To be clear, this is an exercise in algorithmically interrogating a webpage or webserver and getting the desired information. (No credit will be given for manually finding the information and pasting it into a webpage.) Your page should display all statistics displayed upon loading the page, and update them upon the click of a button. It should also display the time at which the data was last retrieved.

Two strategies for obtaining this data are
•	algorithmically interrogating the third-party webpages where this information is already being displayed (e.g. by traversing the DOM) 
•	querying a third-party API where it is available 

There is no specific format for the way the data should be presented, although it should adhere to good UI/UX practices discussed in class last week. Also, the source (URL) of each data should be indicated, somewhere alongside the data itself.